Updated SDCC version to 3.5.5 * Building tests still pending

+++ cpctelera/tools/sdcc-3.5.0/src/src/SDCCopt.c
@@ -1119,7 +1119,7 @@ convertToFcall (eBBlock ** ebbs, int count)
             {
               sym_link *type = operandType (IC_LEFT (ic));
 
-              if (IS_INTEGRAL (type) && getSize (type) > port->support.shift && port->support.shift >= 0)
+              if (IS_INTEGRAL (type) && getSize (type) > (unsigned)port->support.shift && port->support.shift >= 0)
                 {
                   convilong (ic, ebbs[i]);
                 }
